Completed basic functionality
A function was added to the view model to remove all chores marked as done. A delete button was added to the top app bar that calls this function when tapped. Each chore item now includes a checkbox that lets users mark the chore as completed. The view model updates the chore’s status when the checkbox is toggled. A dialog was added to let users type and add new chores. The dialog appears when the add button is pressed and adds the chore to the list when submitted.